Comply with C99, fix modern C errors and enable warnings


See merge request scilab/scilab!837

+++ scilab/modules/hdf5/src/cpp/H5Dataspace.hxx
@@ -63,7 +63,7 @@ public:
             return 0;
         }
 
-        if (H5Sget_simple_extent_ndims(space) != size)
+        if (H5Sget_simple_extent_ndims(space) != (int) size)
         {
             throw H5Exception(__LINE__, __FILE__, _("Invalid selection rank."));
         }
